+++ src/RainBird/responses/CurrentZoneStateResponse.ts
@@ -25,7 +25,19 @@ export class CurrentZoneStateResponse extends Response {
         this._programNumber = undefined;
         this._running = response[3] !== 0;
         break;
-      default: // others such as ESP-ME3
+      case 7: // ESP-ME3 - page 0
+        this._timeRemaining = 0;
+        this._zoneId = 0;
+        this._programNumber = (response[2] <= 3) ? response[2] : 0;
+        this._running = response[3] !== 0;
+        break;
+      case 50: // ESP-ME3 - page 1
+        this._timeRemaining = response.readUInt16BE(4);
+        this._zoneId = response[3];
+        this._programNumber = undefined;
+        this._running = false;
+        break;
+      default: // others
         this._timeRemaining = 0;
         this._zoneId = 0;
         this._programNumber = undefined;